refactor!: move generic equals methods to util package
Moved generic utilities for equality check to util package.

BREAKING CHANGE: moved strictEquals, sameArrayItems, shallowArrayEquals to util package
