+++ packages/wallet/src/services/util/equals.ts
@@ -1,13 +1,7 @@
 import { Cardano, EpochInfo, EraSummary } from '@cardano-sdk/core';
 import { DelegatedStake } from '../types';
 import { GroupedAddress } from '@cardano-sdk/key-management';
-
-export const strictEquals = <T>(a: T, b: T) => a === b;
-
-export const sameArrayItems = <T>(arrayA: T[], arrayB: T[], itemEquals: (a: T, b: T) => boolean) =>
-  arrayA.length === arrayB.length && arrayA.every((a) => arrayB.some((b) => itemEquals(a, b)));
-
-export const shallowArrayEquals = <T>(a: T[], b: T[]) => sameArrayItems(a, b, strictEquals);
+import { sameArrayItems } from '@cardano-sdk/util';
 
 export const tipEquals = (a: Cardano.Tip, b: Cardano.Tip) => a.hash === b.hash;
